refactored Controller\BlogController::class
